+++ js/taxonomy_manager.js
@@ -1,0 +1,6 @@
+var select = document.getElementById('edit-name-select');
+select.onchange = function(){
+  var option = select.options[select.selectedIndex];
+  document.getElementById('edit-addfield').value = option.value;
+};
+